add rpc service
