+++ python/examples/imagenet/get_model.sh
@@ -1,0 +1,2 @@
+wget https://paddle-serving.bj.bcebos.com/imagenet-example/conf_and_model.tar.gz
+tar -xzvf conf_and_model.tar.gz
