+++ python/examples/imagenet/image_reader.py
@@ -12,15 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from paddle_serving_server.web_service import WebService
-import sys
 import cv2
-import base64
 import numpy as np
 
 
-class ImageService(WebService):
-    def set_param(self):
+class ImageReader():
+    def __init__(self):
         self.image_mean = [0.485, 0.456, 0.406]
         self.image_std = [0.229, 0.224, 0.225]
         self.image_shape = [3, 224, 224]
@@ -100,19 +97,3 @@ class ImageService(WebService):
         img -= img_mean
         img /= img_std
         return img
-
-    def preprocess(self, feed={}, fetch=[]):
-        self.set_param()
-        if "image" not in feed:
-            raise ("feed data error!")
-        sample = base64.b64decode(feed["image"])
-        img = self.process_image(sample)
-        res_feed = {}
-        res_feed["image"] = img.reshape(-1)
-        return res_feed, fetch
-
-
-image_service = ImageService(name="image")
-image_service.load_model_config(sys.argv[1])
-image_service.prepare_server(workdir=sys.argv[2], port=9393, device="cpu")
-image_service.run_server()
